The beginings of parted. Much time spend reading D7 code.
The idea behind the new part_of field is that it functions as a
sort of reverse depends_on. It is placed on the child task to ensure
that every task is only ever part of one parent task. For state
purposes, the parent task is considered as depending on its child
tasks, while other properties, like 'project', will be inherited
from the parent to the child.

The parted module is an form element that get added to the node
edit form, which allows for some editing of the task that's part
of the given node. It will give a simplified interface for creating
subtasks, and specifying their description, area and dependencies.
It should also provide a template system for creating templates of
often used task 'patterns'.

It's going to be very ajax heavy.